+++ src/test/java/org/folio/circulation/api/loans/LoanAPITitleTests.java
@@ -18,7 +18,6 @@ import java.util.concurrent.ExecutionException;
 import java.util.concurrent.TimeoutException;
 
 import static org.folio.circulation.api.support.JsonCollectionAssistant.getRecordById;
-import static org.hamcrest.CoreMatchers.nullValue;
 import static org.hamcrest.core.Is.is;
 import static org.hamcrest.junit.MatcherAssert.assertThat;
 
@@ -43,7 +42,6 @@ public class LoanAPITitleTests extends APITests {
 
     UUID itemId = itemsClient.create(
       ItemRequestExamples.basedUponSmallAngryPlanet()
-        .withTitle("A different title") // deliberately different to demonstrate behaviour
         .forHolding(holdingId))
       .getId();
 
@@ -77,50 +75,7 @@ public class LoanAPITitleTests extends APITests {
   }
 
   @Test
-  public void titleIsFromItemWhenNoHolding()
-    throws InterruptedException,
-    ExecutionException,
-    TimeoutException,
-    MalformedURLException,
-    UnsupportedEncodingException {
-
-    UUID itemId = itemsClient.create(
-      ItemRequestExamples.basedUponSmallAngryPlanet()
-        .withTitle("A different title") // deliberately different to demonstrate behaviour
-        .forHolding(null))
-      .getId();
-
-    UUID loanId = UUID.randomUUID();
-
-    IndividualResource response = loansClient.create(new LoanRequestBuilder()
-      .withId(loanId)
-      .withItemId(itemId));
-
-    JsonObject createdLoan = response.getJson();
-
-    assertThat("has item title",
-      createdLoan.getJsonObject("item").containsKey("title"), is(true));
-
-    assertThat("title is taken from item",
-      createdLoan.getJsonObject("item").getString("title"),
-      is("A different title"));
-
-    Response fetchedLoanResponse = loansClient.getById(loanId);
-
-    assertThat(fetchedLoanResponse.getStatusCode(), is(200));
-
-    JsonObject fetchedLoan = fetchedLoanResponse.getJson();
-
-    assertThat("has item title",
-      fetchedLoan.getJsonObject("item").containsKey("title"), is(true));
-
-    assertThat("title is taken from item",
-      fetchedLoan.getJsonObject("item").getString("title"),
-      is("A different title"));
-  }
-
-  @Test
-  public void titleIsFromItemWhenHoldingNotFound()
+  public void noTitleWhenHoldingNotFound()
     throws InterruptedException,
     ExecutionException,
     TimeoutException,
@@ -138,7 +93,6 @@ public class LoanAPITitleTests extends APITests {
 
     UUID itemId = itemsClient.create(
       ItemRequestExamples.basedUponSmallAngryPlanet()
-        .withTitle("A different title") // deliberately different to demonstrate behaviour
         .forHolding(holdingId))
       .getId();
 
@@ -152,12 +106,8 @@ public class LoanAPITitleTests extends APITests {
 
     JsonObject createdLoan = response.getJson();
 
-    assertThat("has item title",
-      createdLoan.getJsonObject("item").containsKey("title"), is(true));
-
-    assertThat("title is taken from item",
-      createdLoan.getJsonObject("item").getString("title"),
-      is("A different title"));
+    assertThat("has no title",
+      createdLoan.getJsonObject("item").containsKey("title"), is(false));
 
     Response fetchedLoanResponse = loansClient.getById(loanId);
 
@@ -165,16 +115,12 @@ public class LoanAPITitleTests extends APITests {
 
     JsonObject fetchedLoan = fetchedLoanResponse.getJson();
 
-    assertThat("has item title",
-      fetchedLoan.getJsonObject("item").containsKey("title"), is(true));
-
-    assertThat("title is taken from item",
-      fetchedLoan.getJsonObject("item").getString("title"),
-      is("A different title"));
+    assertThat("has no title",
+      fetchedLoan.getJsonObject("item").containsKey("title"), is(false));
   }
 
   @Test
-  public void titleIsFromItemWhenInstanceNotFound()
+  public void noTitleWhenInstanceNotFound()
     throws InterruptedException,
     ExecutionException,
     TimeoutException,
@@ -192,7 +138,6 @@ public class LoanAPITitleTests extends APITests {
 
     UUID itemId = itemsClient.create(
       ItemRequestExamples.basedUponSmallAngryPlanet()
-        .withTitle("A different title") // deliberately different to demonstrate behaviour
         .forHolding(holdingId))
       .getId();
 
@@ -206,12 +151,8 @@ public class LoanAPITitleTests extends APITests {
 
     JsonObject createdLoan = response.getJson();
 
-    assertThat("has item title",
-      createdLoan.getJsonObject("item").containsKey("title"), is(true));
-
-    assertThat("title is taken from item",
-      createdLoan.getJsonObject("item").getString("title"),
-      is("A different title"));
+    assertThat("has no title",
+      createdLoan.getJsonObject("item").containsKey("title"), is(false));
 
     Response fetchedLoanResponse = loansClient.getById(loanId);
 
@@ -219,12 +160,8 @@ public class LoanAPITitleTests extends APITests {
 
     JsonObject fetchedLoan = fetchedLoanResponse.getJson();
 
-    assertThat("has item title",
-      fetchedLoan.getJsonObject("item").containsKey("title"), is(true));
-
-    assertThat("title is taken from item",
-      fetchedLoan.getJsonObject("item").getString("title"),
-      is("A different title"));
+    assertThat("has no title",
+      fetchedLoan.getJsonObject("item").containsKey("title"), is(false));
   }
 
   @Test
@@ -245,7 +182,6 @@ public class LoanAPITitleTests extends APITests {
 
     UUID firstItemId = itemsClient.create(
       ItemRequestExamples.basedUponSmallAngryPlanet()
-        .withTitle("A different title") // deliberately different to demonstrate behaviour
         .forHolding(firstHoldingId))
       .getId();
 
@@ -260,7 +196,6 @@ public class LoanAPITitleTests extends APITests {
 
     UUID secondItemId = itemsClient.create(
       ItemRequestExamples.basedUponTemeraire()
-        .withTitle("Another different title") // deliberately different to demonstrate behaviour
         .forHolding(secondHoldingId))
       .getId();
 
@@ -294,7 +229,7 @@ public class LoanAPITitleTests extends APITests {
   }
 
   @Test
-  public void titlesComeFromItemForMultipleLoansWhenHoldingOrInstanceNotFound()
+  public void noTitlesForMultipleLoansWhenHoldingOrInstanceNotFound()
     throws InterruptedException,
     MalformedURLException,
     TimeoutException,
@@ -311,7 +246,6 @@ public class LoanAPITitleTests extends APITests {
 
     UUID firstItemId = itemsClient.create(
       ItemRequestExamples.basedUponSmallAngryPlanet()
-        .withTitle("A different title") // deliberately different to demonstrate behaviour
         .forHolding(firstHoldingId))
       .getId();
 
@@ -326,7 +260,6 @@ public class LoanAPITitleTests extends APITests {
 
     UUID secondItemId = itemsClient.create(
       ItemRequestExamples.basedUponTemeraire()
-        .withTitle("Another different title") // deliberately different to demonstrate behaviour
         .forHolding(secondHoldingId))
       .getId();
 
@@ -349,75 +282,10 @@ public class LoanAPITitleTests extends APITests {
     JsonObject secondFetchedLoan = getRecordById(
       fetchedLoansResponse, secondLoanId).get();
 
-    assertThat("has item title",
-      firstFetchedLoan.getJsonObject("item").containsKey("title"), is(true));
+    assertThat("has no title",
+      firstFetchedLoan.getJsonObject("item").containsKey("title"), is(false));
 
-    assertThat("title is taken from item",
-      firstFetchedLoan.getJsonObject("item").getString("title"),
-      is("A different title"));
-
-    assertThat("has item title",
-      secondFetchedLoan.getJsonObject("item").containsKey("title"), is(true));
-
-    assertThat("title is taken from item",
-      secondFetchedLoan.getJsonObject("item").getString("title"),
-      is("Another different title"));
-  }
-
-  @Test
-  public void titlesComeFromItemForMultipleLoansWhenNoHolding()
-    throws InterruptedException,
-    MalformedURLException,
-    TimeoutException,
-    ExecutionException {
-
-    UUID firstItemId = itemsClient.create(
-      ItemRequestExamples.basedUponSmallAngryPlanet()
-        .withTitle("A different title")
-        .forHolding(null))
-      .getId();
-
-    UUID firstLoanId = loansClient.create(new LoanRequestBuilder()
-      .withItemId(firstItemId)).getId();
-
-    List<JsonObject> fetchedLoansResponse = loansClient.getAll();
-
-    JsonObject firstFetchedLoan = getRecordById(
-      fetchedLoansResponse, firstLoanId).get();
-
-    assertThat("has item title",
-      firstFetchedLoan.getJsonObject("item").containsKey("title"), is(true));
-
-    assertThat("title is taken from item",
-      firstFetchedLoan.getJsonObject("item").getString("title"),
-      is("A different title"));
-  }
-
-  @Test
-  public void noTitleWhenForMultipleLoansWhenItemHasNoTitleAndNoHolding()
-    throws InterruptedException,
-    MalformedURLException,
-    TimeoutException,
-    ExecutionException {
-
-    UUID firstItemId = itemsClient.create(
-      ItemRequestExamples.basedUponSmallAngryPlanet()
-        .withNoTitle()
-        .forHolding(null))
-      .getId();
-
-    UUID firstLoanId = loansClient.create(new LoanRequestBuilder()
-      .withItemId(firstItemId)).getId();
-
-    List<JsonObject> fetchedLoansResponse = loansClient.getAll();
-
-    JsonObject firstFetchedLoan = getRecordById(
-      fetchedLoansResponse, firstLoanId).get();
-
-    assertThat("has item title",
-      firstFetchedLoan.getJsonObject("item").containsKey("title"), is(true));
-
-    assertThat("has blank item title",
-      firstFetchedLoan.getJsonObject("item").getString("title"), is(nullValue()));
+    assertThat("has no title",
+      secondFetchedLoan.getJsonObject("item").containsKey("title"), is(false));
   }
 }