+++ src/test/java/org/folio/circulation/api/requests/RequestsAPITitleTests.java
@@ -44,7 +44,6 @@ public class RequestsAPITitleTests extends APITests {
 
     UUID itemId = itemsClient.create(
       ItemRequestExamples.basedUponSmallAngryPlanet()
-        .withTitle("A different title on the item") // deliberately different to demonstrate behaviour
         .forHolding(holdingId))
       .getId();
 
@@ -81,7 +80,7 @@ public class RequestsAPITitleTests extends APITests {
   }
 
   @Test
-  public void titleIsFromItemWhenCreatingRequestWithNoHolding()
+  public void noTitleWhenCreatingRequestForNotFoundHolding()
     throws InterruptedException,
     ExecutionException,
     TimeoutException,
@@ -99,7 +98,6 @@ public class RequestsAPITitleTests extends APITests {
 
     UUID itemId = itemsClient.create(
       ItemRequestExamples.basedUponSmallAngryPlanet()
-        .withTitle("A different title on the item") // deliberately different to demonstrate behaviour
         .forHolding(holdingId))
       .getId();
 
@@ -116,12 +114,8 @@ public class RequestsAPITitleTests extends APITests {
 
     JsonObject createdRequest = response.getJson();
 
-    assertThat("has item title",
-      createdRequest.getJsonObject("item").containsKey("title"), is(true));
-
-    assertThat("title is taken from item",
-      createdRequest.getJsonObject("item").getString("title"),
-      is("A different title on the item"));
+    assertThat("has no title",
+      createdRequest.getJsonObject("item").containsKey("title"), is(false));
 
     Response fetchedRequestResponse = requestsClient.getById(requestId);
 
@@ -129,16 +123,12 @@ public class RequestsAPITitleTests extends APITests {
 
     JsonObject fetchedRequest = fetchedRequestResponse.getJson();
 
-    assertThat("has item title",
-      fetchedRequest.getJsonObject("item").containsKey("title"), is(true));
-
-    assertThat("title is taken from item",
-      fetchedRequest.getJsonObject("item").getString("title"),
-      is("A different title on the item"));
+    assertThat("has no title",
+      fetchedRequest.getJsonObject("item").containsKey("title"), is(false));
   }
 
   @Test
-  public void titleIsFromItemWhenCreatingRequestWithNoInstance()
+  public void noTitleWhenCreatingRequestAndInstanceNotFound()
     throws InterruptedException,
     ExecutionException,
     TimeoutException,
@@ -156,7 +146,6 @@ public class RequestsAPITitleTests extends APITests {
 
     UUID itemId = itemsClient.create(
       ItemRequestExamples.basedUponSmallAngryPlanet()
-        .withTitle("A different title on the item") // deliberately different to demonstrate behaviour
         .forHolding(holdingId))
       .getId();
 
@@ -173,12 +162,8 @@ public class RequestsAPITitleTests extends APITests {
 
     JsonObject createdRequest = response.getJson();
 
-    assertThat("has item title",
-      createdRequest.getJsonObject("item").containsKey("title"), is(true));
-
-    assertThat("title is taken from item",
-      createdRequest.getJsonObject("item").getString("title"),
-      is("A different title on the item"));
+    assertThat("has no title",
+      createdRequest.getJsonObject("item").containsKey("title"), is(false));
 
     Response fetchedRequestResponse = requestsClient.getById(requestId);
 
@@ -186,12 +171,8 @@ public class RequestsAPITitleTests extends APITests {
 
     JsonObject fetchedRequest = fetchedRequestResponse.getJson();
 
-    assertThat("has item title",
-      fetchedRequest.getJsonObject("item").containsKey("title"), is(true));
-
-    assertThat("title is taken from item",
-      fetchedRequest.getJsonObject("item").getString("title"),
-      is("A different title on the item"));
+    assertThat("has no title",
+      fetchedRequest.getJsonObject("item").containsKey("title"), is(false));
   }
 
   @Test
@@ -213,7 +194,6 @@ public class RequestsAPITitleTests extends APITests {
 
     UUID itemId = itemsClient.create(
       ItemRequestExamples.basedUponSmallAngryPlanet()
-        .withTitle("A different title") // deliberately different to demonstrate behaviour
         .forHolding(holdingId))
       .getId();
 
@@ -250,7 +230,7 @@ public class RequestsAPITitleTests extends APITests {
   }
 
   @Test
-  public void titleIsChangedWhenRequestUpdatedAndNoInstance()
+  public void noTitleWhenRequestUpdatedAndInstanceNotFound()
     throws InterruptedException,
     ExecutionException,
     TimeoutException,
@@ -268,7 +248,6 @@ public class RequestsAPITitleTests extends APITests {
 
     UUID itemId = itemsClient.create(
       ItemRequestExamples.basedUponSmallAngryPlanet()
-        .withTitle("A different title on item") // deliberately different to demonstrate behaviour
         .forHolding(holdingId))
       .getId();
 
@@ -293,57 +272,12 @@ public class RequestsAPITitleTests extends APITests {
 
     JsonObject fetchedRequest = fetchedRequestResponse.getJson();
 
-    assertThat("has item title",
-      fetchedRequest.getJsonObject("item").containsKey("title"), is(true));
-
-    assertThat("title is taken from item",
-      fetchedRequest.getJsonObject("item").getString("title"),
-      is("A different title on item"));
+    assertThat("has no title",
+      fetchedRequest.getJsonObject("item").containsKey("title"), is(false));
   }
 
   @Test
-  public void titleIsChangedWhenRequestUpdatedAndNoHolding()
-    throws InterruptedException,
-    ExecutionException,
-    TimeoutException,
-    MalformedURLException,
-    UnsupportedEncodingException {
-
-    UUID itemId = itemsClient.create(
-      ItemRequestExamples.basedUponSmallAngryPlanet()
-        .withTitle("A different title on item") // deliberately different to demonstrate behaviour
-        .forHolding(null))
-      .getId();
-
-    checkOutItem(itemId, loansClient);
-
-    UUID requestId = UUID.randomUUID();
-
-    IndividualResource response = requestsClient.create(new RequestRequestBuilder()
-      .withId(requestId)
-      .withRequesterId(usersClient.create(new UserRequestBuilder().create()).getId())
-      .withItemId(itemId));
-
-    JsonObject createdRequest = response.getJson();
-
-    requestsClient.replace(requestId, createdRequest);
-
-    Response fetchedRequestResponse = requestsClient.getById(requestId);
-
-    assertThat(fetchedRequestResponse.getStatusCode(), is(200));
-
-    JsonObject fetchedRequest = fetchedRequestResponse.getJson();
-
-    assertThat("has item title",
-      fetchedRequest.getJsonObject("item").containsKey("title"), is(true));
-
-    assertThat("title is taken from item",
-      fetchedRequest.getJsonObject("item").getString("title"),
-      is("A different title on item"));
-  }
-
-  @Test
-  public void titleIsChangedWhenRequestUpdatedAndHoldingNotFound()
+  public void noTitleWhenRequestUpdatedAndHoldingNotFound()
     throws InterruptedException,
     ExecutionException,
     TimeoutException,
@@ -361,7 +295,6 @@ public class RequestsAPITitleTests extends APITests {
 
     UUID itemId = itemsClient.create(
       ItemRequestExamples.basedUponSmallAngryPlanet()
-        .withTitle("A different title on item") // deliberately different to demonstrate behaviour
         .forHolding(holdingId))
       .getId();
 
@@ -386,12 +319,8 @@ public class RequestsAPITitleTests extends APITests {
 
     JsonObject fetchedRequest = fetchedRequestResponse.getJson();
 
-    assertThat("has item title",
-      fetchedRequest.getJsonObject("item").containsKey("title"), is(true));
-
-    assertThat("title is taken from item",
-      fetchedRequest.getJsonObject("item").getString("title"),
-      is("A different title on item"));
+    assertThat("has no title",
+      fetchedRequest.getJsonObject("item").containsKey("title"), is(false));
   }
 
   @Test
@@ -412,7 +341,6 @@ public class RequestsAPITitleTests extends APITests {
 
     UUID firstItemId = itemsClient.create(
       ItemRequestExamples.basedUponSmallAngryPlanet()
-        .withTitle("A different title") // deliberately different to demonstrate behaviour
         .forHolding(firstHoldingId))
       .getId();
 
@@ -427,7 +355,6 @@ public class RequestsAPITitleTests extends APITests {
 
     UUID secondItemId = itemsClient.create(
       ItemRequestExamples.basedUponTemeraire()
-        .withTitle("Another different title") // deliberately different to demonstrate behaviour
         .forHolding(secondHoldingId))
       .getId();
 
@@ -486,7 +413,6 @@ public class RequestsAPITitleTests extends APITests {
 
     UUID firstItemId = itemsClient.create(
       ItemRequestExamples.basedUponSmallAngryPlanet()
-        .withTitle("A different title") // deliberately different to demonstrate behaviour
         .forHolding(firstHoldingId))
       .getId();
 
@@ -501,7 +427,6 @@ public class RequestsAPITitleTests extends APITests {
 
     UUID secondItemId = itemsClient.create(
       ItemRequestExamples.basedUponTemeraire()
-        .withTitle("Another different title") // deliberately different to demonstrate behaviour
         .forHolding(secondHoldingId))
       .getId();